refactor(components): use IconFC in components
