+++ packages/components/src/components/tree-item/component.tsx
@@ -1,11 +1,11 @@
 import { Component, Element, h, Host, type JSX, Method, Prop, State, Watch } from '@stencil/core';
 
 import { KolLinkWcTag, KolTreeTag } from '../../core/component-names';
-import IconFC from '../../functional-components/Icon';
 import type { ActivePropType, HrefPropType, LabelPropType, OpenPropType, TreeItemAPI, TreeItemStates } from '../../schema';
 import { validateActive, validateHref, validateLabel, validateOpen } from '../../schema';
 import clsx from '../../utils/clsx';
 import { nonce } from '../../utils/dev.utils';
+import { IconFC } from '../_skeleton/internal/functional-components/icon/component';
 
 @Component({
 	tag: `kol-tree-item-wc`,
