+++ packages/components/src/functional-components/index.ts
@@ -3,7 +3,6 @@ export { default as KolCollapsibleFc } from './Collapsible';
 export { default as KolFormFieldFc } from './FormField';
 export { default as KolFormFieldMsgFc } from './FormFieldMsg';
 export { default as KolHeadingFc } from './Heading';
-export { default as KolIconFc } from './Icon';
 export { default as KolInputContainerFc } from './InputContainer';
 export { default as KolInputFc } from './inputs/Input';
 export { default as KolTextareaFc } from './inputs/TextArea';
